refactor(JanggiPosition): JanggiPositionFactory를 JanggiPosition 내부로 변경
